+++ apps/backend/prisma/migrations/20240904155848_add_bitbucket_oauth/migration.sql
@@ -1,0 +1,2 @@
+-- AlterEnum
+ALTER TYPE "StandardOAuthProviderType" ADD VALUE 'BITBUCKET';