+++ apps/backend/prisma/migrations/20240905201445_ms_tenant/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "StandardOAuthProviderConfig" ADD COLUMN     "microsoftTenantId" TEXT;