+++ packages/stack/src/components-page/sign-in.tsx
@@ -1,5 +1,5 @@
 import { AuthPage } from './auth-page';
 
-export function SignIn({ fullPage=false }: { fullPage?: boolean }) {
-  return <AuthPage fullPage={fullPage} type='sign-in' />;
+export function SignIn(props: { fullPage?: boolean, automaticRedirect?: boolean, extraInfo?: React.ReactNode }) {
+  return <AuthPage fullPage={!!props.fullPage} type='sign-in' automaticRedirect={!!props.automaticRedirect} extraInfo={props.extraInfo} />;
 }