+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stackframe/stack-ui
 
+## 2.5.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.30
+
+## 2.5.29
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.5.29
+
 ## 2.5.28
 
 ### Patch Changes